joshua-holtsclaw

+++ data/migrations/20191116140506_totalTable.js
@@ -33,7 +33,7 @@ exports.up = function(knex) {
         tbl.string('THC_Percent',255).notNullable();
         tbl.string('CBD',255).notNullable();
         tbl.string('Description1',255).notNullable();
-        tbl.integer('Score').notNullable();
+        tbl.float('Score').notNullable();
         tbl.integer('Recommendation');
         tbl.integer('exp');
         tbl.integer('iat');
